Move validator-specific tests to validator pkg
